Refactored bgs_errors.rb and bgs.rb
