--- app/exceptions/bgs_errors.rb
@@ -1,8 +1,0 @@
-module BGS
-  class InvalidUsername < StandardError; end
-  class InvalidStation < StandardError; end
-  class InvalidApplication < StandardError; end
-  class NoActiveStations < StandardError; end
-  class NoCaseflowAccess < StandardError; end
-  class StationAssertionRequired < StandardError; end
-end
