+++ config/initializers/bgs.rb
@@ -1,3 +1,8 @@
-Rails.application.reloader.to_prepare do
-  BGSService = (!BaseController.dependencies_faked? ? ExternalApi::BGSService : Fakes::BGSService)
+module BGS
+  class InvalidUsername < StandardError; end
+  class InvalidStation < StandardError; end
+  class InvalidApplication < StandardError; end
+  class NoActiveStations < StandardError; end
+  class NoCaseflowAccess < StandardError; end
+  class StationAssertionRequired < StandardError; end
 end
