+++ config/initializers/bgs_service.rb
@@ -1,0 +1,3 @@
+Rails.application.reloader.to_prepare do
+  BGSService = (!BaseController.dependencies_faked? ? ExternalApi::BGSService : Fakes::BGSService)
+end
